requirements.txt
aiohttp>=3.8.5
azure-ai-ml==1.23.0
azure-identity==1.15.0
azure-ml==0.0.1
azure-ml-component==0.9.18.post2
azureml-contrib-services==1.56.0
azureml-core==1.56.0
azureml-dataset-runtime==1.56.0
azureml-defaults==1.56.0.post1
azureml-inference-server-http
azureml-mlflow==1.56.0
captum==0.8.0
cryptography>=42.0.5
debugpy>=1.8.0
gdcm==1.1
idna>=3.7
inference-schema
ipykernel>=6.0
lungmask
MarkupSafe==2.1.2
matplotlib>=3.8.0
nibabel>=5.0.0
numpy>=1.26.0
opencv-python>=4.8.0
Pillow
psutil>=5.8.0
py-cpuinfo==5.0.0
py-spy==0.3.12
pybind11
pydicom==2.4.4
# pydicom==1.4.2
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.2.0
pylibjpeg-openjpeg==2.3.0
regex
rich
scikit-learn>=1.3.2
scipy>=1.11.1
seaborn==0.13.2
tabulate==0.9.0
tensorboard
timm==1.0.12
torch==2.3.1
torch-tb-profiler>=0.4.0
TotalSegmentator
tqdm>=4.66.3
transformers>=4.0.0
unet_lungs_segmentation
urllib3>=1.26.18
torchsummary==1.5.1
torchinfo==1.8.0